refactor(app): move stream reducer into timeline
The reducer is pure timeline and stream state logic, so keeping it with timeline code makes the pagination path easier to review without changing behavior.
